feat: D5 de-god-module — split consciousness.py into 8 cohesive sub-modules
consciousness.py (3,277 lines, 41 MCP tools) split via strangler fig into:
- n8n.py (82L, 2 tools) — webhook dispatch
- emotion.py (441L, 9 tools) — PAD model, mood, emotion-tagged memories
- prediction.py (270L, 4 tools) — predict, surprise, beliefs, accuracy
- workspace.py (461L, 6 tools) — GWT spotlight, attention, salience
- learning.py (280L, 3 tools) — auto-learn, tool audit, topic confidence
- self_model.py (756L, 5 tools) — reflect, assess, gaps, butlin
- curiosity.py (298L, 5 tools) — curiosidad, patrones, novedad
- lifecycle.py (845L, 7 tools) — despertar, salud, ciclo_vida, consolidation

consciousness.py becomes a 94-line facade with eager re-exports + lazy
__getattr__ for lifecycle. Back-compat re-exports (qdrant, memory,
_classify_emotion, _get_emotion_text) ensure monkeypatching in tests works.

Contract tests added: back-compat names, identity refs, lazy lifecycle.
373 tests pass, 0 regressions. No circular imports. 3/3 identity checks OK.
