+++ tests/test_continuity_battery.py
@@ -1,0 +1,838 @@
+"""
+CONTINUITY BATTERY TESTS
+=========================
+23 tests across 9 buckets validating the cross-session continuity pipeline end-to-end.
+
+Each test validates a neuroscientifically grounded mechanism, not just absence of crashes.
+
+Buckets:
+  1. Hippocampal Replay -- Checkpoint Fidelity (3)
+  2. Pattern Completion -- Bridge Reconstruction (2)
+  3. Cross-Session Prediction Error (5)
+  4. Synaptic Homeostasis (3)
+  5. Prospective Memory Persistence (3)
+  6. Autobiographical Continuity (2)
+  7. State-Dependent Retrieval (2)
+  8. Sleep Loop Mechanics (2)
+  9. Source Priority (1)
+
+Run: ./venv/bin/pytest tests/test_continuity_battery.py -v
+"""
+
+import sys
+import os
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+
+import json
+import random
+import sqlite3
+import pytest
+from datetime import datetime, timedelta
+from unittest.mock import MagicMock
+from zoneinfo import ZoneInfo
+
+TZ_COL = ZoneInfo("America/Bogota")
+
+pytestmark = pytest.mark.continuity
+
+PROJECT_ROOT = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+FTS_MIGRATIONS_DIR = os.path.join(PROJECT_ROOT, "migrations")
+
+
+# ============================================================
+# SHARED FIXTURES
+# ============================================================
+
+@pytest.fixture
+def continuity_db(tmp_path, monkeypatch):
+    """Isolated SQLite DB with full migrations + WM tables."""
+    db_path = str(tmp_path / "test_fts.db")
+    monkeypatch.setenv("FTS_DB_PATH", db_path)
+    monkeypatch.setattr("modules.config.FTS_DB_PATH", db_path)
+
+    from modules.migrations import apply_migrations
+    apply_migrations(db_path, migrations_dir=FTS_MIGRATIONS_DIR)
+
+    conn = sqlite3.connect(db_path)
+    conn.execute("PRAGMA journal_mode=WAL")
+    conn.executescript("""
+        CREATE TABLE IF NOT EXISTS working_memory (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            content TEXT NOT NULL,
+            topic TEXT DEFAULT 'general',
+            relevance REAL DEFAULT 0.5,
+            added_at TEXT NOT NULL,
+            occurred_at TEXT,
+            source TEXT DEFAULT 'interaction',
+            related_memory_id TEXT,
+            chain_id TEXT,
+            active INTEGER DEFAULT 1,
+            last_accessed_at TEXT,
+            access_count INTEGER DEFAULT 0
+        );
+        CREATE TABLE IF NOT EXISTS narrative_traces (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            trace_name TEXT NOT NULL UNIQUE,
+            chain_ids TEXT NOT NULL,
+            theme TEXT,
+            created_at TEXT NOT NULL,
+            last_updated TEXT NOT NULL,
+            active INTEGER DEFAULT 1
+        );
+    """)
+    conn.commit()
+    conn.close()
+
+    monkeypatch.setattr("modules.session_bridge.FTS_DB_PATH", db_path)
+    monkeypatch.setattr("modules.sleep_loop.FTS_DB_PATH", db_path)
+    monkeypatch.setattr("modules.sleep_loop.DATA_DIR", str(tmp_path))
+    monkeypatch.setattr("modules.sleep_loop.LOCK_FILE", str(tmp_path / "sleep_loop.lock"))
+    monkeypatch.setattr("modules.working_memory.FTS_DB_PATH", db_path)
+    monkeypatch.setattr("modules.working_memory._TABLES_INITIALIZED", False)
+
+    return db_path
+
+
+@pytest.fixture
+def mock_external_deps(monkeypatch):
+    """Mock 7+ external sources that checkpoint_session_close collects from."""
+    monkeypatch.setattr("modules.wiring._attention_schema", {
+        "current_focus": "continuity_test",
+        "focus_strength": 0.75,
+    })
+    monkeypatch.setattr("modules.session_bridge._emotional_state", {
+        "current": {
+            "pleasure": 0.6, "arousal": -0.3, "dominance": 0.7,
+            "trigger": "bucket_test",
+        },
+        "mood": {"pleasure": 0.2, "arousal": 0.1, "dominance": 0.3, "last_updated": None},
+        "history": [],
+        "decay_rate": 0.1,
+        "mood_shift_rate": 0.05,
+    })
+    monkeypatch.setattr(
+        "modules.prospective.get_pending_intentions",
+        lambda limit=10: []
+    )
+    monkeypatch.setattr("modules.session_bridge.get_trace_id", lambda: "bat_trace_001")
+    monkeypatch.setattr("modules.session_bridge._current_session", "bat-session-001")
+
+
+@pytest.fixture
+def clean_pad():
+    """Save/restore _emotional_state to prevent PAD leaks between tests."""
+    from modules.config import _emotional_state
+    import copy
+    saved = copy.deepcopy(_emotional_state)
+    yield _emotional_state
+    _emotional_state.clear()
+    _emotional_state.update(saved)
+
+
+def _insert_checkpoint(db_path, **overrides):
+    """Helper to insert a checkpoint with optional overrides."""
+    from modules.config import now_col as _now_col
+    defaults = {
+        "session_summary": "test session",
+        "source": "flush",
+        "created_at": _now_col().isoformat(),
+        "sleep_report": None,
+        "pad_pleasure": 0.0,
+        "pad_arousal": 0.0,
+        "pad_dominance": 0.0,
+        "pad_trigger": "",
+        "pending_intentions": "[]",
+        "goal_stack": "[]",
+        "active_project": None,
+        "wm_items": "[]",
+        "wm_active_count": 0,
+        "last_decisions": "[]",
+        "active_traces": "[]",
+    }
+    defaults.update(overrides)
+    conn = sqlite3.connect(db_path)
+    cols = list(defaults.keys())
+    placeholders = ", ".join(["?"] * len(cols))
+    conn.execute(
+        f"INSERT INTO session_checkpoints ({', '.join(cols)}) VALUES ({placeholders})",
+        list(defaults.values())
+    )
+    conn.commit()
+    last_id = conn.execute("SELECT last_insert_rowid()").fetchone()[0]
+    conn.close()
+    return last_id
+
+
+# ============================================================
+# FAKE QDRANT (for Bucket 4)
+# ============================================================
+
+class FakeQdrantPoint:
+    """Minimal point mimic for Qdrant scroll results."""
+    def __init__(self, point_id, payload):
+        self.id = point_id
+        self.payload = payload
+
+
+class FakeQdrant:
+    """Minimal Qdrant fake for homeostasis tests."""
+    def __init__(self, points):
+        self.points = points
+        self.payloads_set = []
+
+    def scroll(self, collection_name, scroll_filter=None, limit=100,
+               with_payload=True, with_vectors=False, offset=None):
+        return (self.points, None)
+
+    def set_payload(self, collection_name, payload, points):
+        self.payloads_set.append({"payload": payload, "points": points})
+
+
+# ============================================================
+# BUCKET 1: Hippocampal Replay -- Checkpoint Fidelity
+# Ref: Wilson & McNaughton 1994, Diekelmann & Born 2010
+# ============================================================
+
+class TestCheckpointFidelity:
+    """Bucket 1: checkpoint_session_close captures runtime state faithfully."""
+
+    def test_checkpoint_captures_pad_state(self, continuity_db, mock_external_deps):
+        """PAD (0.6, -0.3, 0.7) + trigger in DB == runtime state."""
+        from modules.session_bridge import checkpoint_session_close
+
+        result = checkpoint_session_close(
+            session_summary="PAD fidelity test",
+            source="flush",
+        )
+        assert result["ok"] is True
+
+        conn = sqlite3.connect(continuity_db)
+        row = conn.execute(
+            "SELECT pad_pleasure, pad_arousal, pad_dominance, pad_trigger "
+            "FROM session_checkpoints WHERE id = ?",
+            (result["checkpoint_id"],)
+        ).fetchone()
+        conn.close()
+
+        assert row is not None
+        assert abs(row[0] - 0.6) < 0.01, f"pleasure mismatch: {row[0]}"
+        assert abs(row[1] - (-0.3)) < 0.01, f"arousal mismatch: {row[1]}"
+        assert abs(row[2] - 0.7) < 0.01, f"dominance mismatch: {row[2]}"
+        assert row[3] == "bucket_test"
+
+    def test_checkpoint_captures_pending_intentions(self, continuity_db, monkeypatch, mock_external_deps):
+        """2 intentions JSON-serialized faithfully (priority, expiry null)."""
+        from modules.session_bridge import checkpoint_session_close
+
+        fake_intentions = [
+            {"id": "i1", "action": "Deploy trading v2", "priority": "high",
+             "activation": 0.8, "expiry": None},
+            {"id": "i2", "action": "Review session bridge health", "priority": "critical",
+             "activation": 0.95, "expiry": "2026-03-01T10:00:00-05:00"},
+        ]
+        monkeypatch.setattr(
+            "modules.prospective.get_pending_intentions",
+            lambda limit=10: fake_intentions
+        )
+
+        result = checkpoint_session_close(session_summary="intentions test", source="flush")
+        assert result["ok"] is True
+
+        conn = sqlite3.connect(continuity_db)
+        row = conn.execute(
+            "SELECT pending_intentions FROM session_checkpoints WHERE id = ?",
+            (result["checkpoint_id"],)
+        ).fetchone()
+        conn.close()
+
+        stored = json.loads(row[0])
+        assert len(stored) == 2
+        assert stored[0]["priority"] == "high"
+        assert stored[0]["expiry"] is None
+        assert stored[1]["priority"] == "critical"
+        assert "2026-03-01" in stored[1]["expiry"]
+
+    def test_checkpoint_captures_wm_items(self, continuity_db, mock_external_deps):
+        """Top 15 WM items, content truncated to 200 chars, wm_active_count correct."""
+        from modules.session_bridge import checkpoint_session_close
+        from modules.config import now_col as _now_col
+
+        # Insert 18 WM items (should only capture 15)
+        conn = sqlite3.connect(continuity_db)
+        now = _now_col().isoformat()
+        for i in range(18):
+            content = f"WM item {i}: " + "x" * 250  # >200 chars
+            conn.execute(
+                "INSERT INTO working_memory (content, topic, relevance, added_at, active) "
+                "VALUES (?, 'test', ?, ?, 1)",
+                (content, 0.9 - i * 0.01, now)
+            )
+        conn.commit()
+        conn.close()
+
+        result = checkpoint_session_close(session_summary="wm test", source="flush")
+        assert result["ok"] is True
+
+        conn = sqlite3.connect(continuity_db)
+        row = conn.execute(
+            "SELECT wm_items, wm_active_count FROM session_checkpoints WHERE id = ?",
+            (result["checkpoint_id"],)
+        ).fetchone()
+        conn.close()
+
+        items = json.loads(row[0])
+        assert len(items) <= 15, f"Expected max 15 items, got {len(items)}"
+        assert row[1] == 18, f"Expected wm_active_count=18, got {row[1]}"
+        # Content should be truncated to 200
+        for item in items:
+            assert len(item["content"]) <= 200
+
+
+# ============================================================
+# BUCKET 2: Pattern Completion -- Bridge Reconstruction
+# Ref: Marr 1971, McClelland et al. 1995
+# ============================================================
+
+class TestBridgeReconstruction:
+    """Bucket 2: load_session_bridge reconstructs a grounded narrative."""
+
+    def test_bridge_text_grounded_in_checkpoint_data(self, continuity_db, mock_external_deps):
+        """Summary and project in bridge. No [PE], [Sleep], 'Siguiente paso' when no data."""
+        from modules.session_bridge import checkpoint_session_close, load_session_bridge
+
+        checkpoint_session_close(
+            session_summary="Implementamos session bridge v1",
+            source="flush",
+        )
+
+        bridge = load_session_bridge()
+        assert bridge is not None
+        text = bridge["bridge_text"]
+
+        # Grounded content present
+        assert "Implementamos session bridge v1" in text
+        assert "Han pasado" in text
+
+        # No hallucinated content
+        assert "[PE]" not in text
+        assert "[Sleep]" not in text
+        assert "Siguiente paso" not in text
+
+    def test_bridge_hours_since_last_accuracy(self, continuity_db, mock_external_deps, monkeypatch):
+        """hours_since_last == 3.0 +/- 0.15h for checkpoint 3h ago."""
+        from modules.session_bridge import load_session_bridge
+        from modules.config import now_col as _now_col
+
+        three_hours_ago = (_now_col() - timedelta(hours=3)).isoformat()
+        _insert_checkpoint(continuity_db, created_at=three_hours_ago,
+                           session_summary="3 hours ago session")
+
+        bridge = load_session_bridge()
+        assert bridge is not None
+        assert abs(bridge["hours_since_last"] - 3.0) < 0.15, \
+            f"Expected ~3.0h, got {bridge['hours_since_last']}"
+
+
+# ============================================================
+# BUCKET 3: Cross-Session Prediction Error
+# Ref: Friston 2010, Schultz 1997
+# ============================================================
+
+class TestCrossSessionPE:
+    """Bucket 3: deterministic prediction error generation from checkpoint data."""
+
+    def test_expired_intention_generates_pe(self, continuity_db, mock_external_deps):
+        """Intention with expiry in the past -> 1 PE type='intention_expired'."""
+        from modules.session_bridge import load_session_bridge
+        from modules.config import now_col as _now_col
+
+        past_expiry = (_now_col() - timedelta(hours=2)).isoformat()
+        intentions = json.dumps([{
+            "id": "exp1", "action": "Deploy before deadline",
+            "priority": "high", "activation": 0.7, "expiry": past_expiry,
+        }])
+        _insert_checkpoint(continuity_db, pending_intentions=intentions,
+                           session_summary="expired intention test")
+
+        bridge = load_session_bridge()
+        assert bridge is not None
+        pes = bridge["prediction_errors"]
+        assert len(pes) == 1
+        assert pes[0]["type"] == "intention_expired"
+        assert "Deploy before deadline" in pes[0]["detail"]
+
+    def test_non_expired_intention_no_pe(self, continuity_db, mock_external_deps):
+        """Intention with expiry in the future -> 0 PEs."""
+        from modules.session_bridge import load_session_bridge
+        from modules.config import now_col as _now_col
+
+        future_expiry = (_now_col() + timedelta(hours=24)).isoformat()
+        intentions = json.dumps([{
+            "id": "fut1", "action": "Future task",
+            "priority": "medium", "activation": 0.5, "expiry": future_expiry,
+        }])
+        _insert_checkpoint(continuity_db, pending_intentions=intentions,
+                           session_summary="future intention test")
+
+        bridge = load_session_bridge()
+        assert bridge is not None
+        assert len(bridge["prediction_errors"]) == 0
+
+    def test_intention_without_expiry_no_pe(self, continuity_db, mock_external_deps):
+        """Intention with expiry=None -> 0 PEs."""
+        from modules.session_bridge import load_session_bridge
+
+        intentions = json.dumps([{
+            "id": "nul1", "action": "Open-ended task",
+            "priority": "low", "activation": 0.3, "expiry": None,
+        }])
+        _insert_checkpoint(continuity_db, pending_intentions=intentions,
+                           session_summary="null expiry test")
+
+        bridge = load_session_bridge()
+        assert bridge is not None
+        assert len(bridge["prediction_errors"]) == 0
+
+    def test_external_event_newer_generates_pe(self, continuity_db, mock_external_deps, tmp_path, monkeypatch):
+        """Recordatorio with timestamp > checkpoint -> PE type='external_event'."""
+        from modules.session_bridge import load_session_bridge
+        from modules.config import now_col as _now_col
+
+        # Checkpoint created 2h ago
+        cp_time = (_now_col() - timedelta(hours=2)).isoformat()
+        _insert_checkpoint(continuity_db, created_at=cp_time,
+                           session_summary="external event test")
+
+        # Recordatorio 1h ago (newer than checkpoint)
+        rec_time = (_now_col() - timedelta(hours=1)).isoformat()
+        rec_file = tmp_path / "rec.json"
+        rec_file.write_text(json.dumps({
+            "pendientes": [{
+                "mensaje": "WhatsApp: nuevo pedido urgente",
+                "timestamp": rec_time,
+                "prioridad": "high",
+            }]
+        }))
+        monkeypatch.setattr("modules.session_bridge.REMINDERS_FILE", str(rec_file))
+
+        bridge = load_session_bridge()
+        assert bridge is not None
+        pes = bridge["prediction_errors"]
+        assert len(pes) == 1
+        assert pes[0]["type"] == "external_event"
+        assert "pedido urgente" in pes[0]["detail"]
+
+    def test_external_event_older_no_pe(self, continuity_db, mock_external_deps, tmp_path, monkeypatch):
+        """Recordatorio with timestamp < checkpoint -> 0 PEs."""
+        from modules.session_bridge import load_session_bridge
+        from modules.config import now_col as _now_col
+
+        # Checkpoint created now
+        _insert_checkpoint(continuity_db, session_summary="old event test")
+
+        # Recordatorio from 3h ago (older than checkpoint)
+        old_time = (_now_col() - timedelta(hours=3)).isoformat()
+        rec_file = tmp_path / "rec.json"
+        rec_file.write_text(json.dumps({
+            "pendientes": [{
+                "mensaje": "Old reminder",
+                "timestamp": old_time,
+                "prioridad": "normal",
+            }]
+        }))
+        monkeypatch.setattr("modules.session_bridge.REMINDERS_FILE", str(rec_file))
+
+        bridge = load_session_bridge()
+        assert bridge is not None
+        assert len(bridge["prediction_errors"]) == 0
+
+
+# ============================================================
+# BUCKET 4: Synaptic Homeostasis
+# Ref: Tononi & Cirelli 2003/2006 (SHY)
+# ============================================================
+
+class TestSynapticHomeostasis:
+    """Bucket 4: salience decay + emotional decay behave correctly."""
+
+    def test_salience_decay_reduces_medium(self, monkeypatch):
+        """apply_salience_decay() reduces salience of medium-importance memories."""
+        points = [
+            FakeQdrantPoint("m1", {"attention_salience": 0.6, "narrative_importance": "medium"}),
+            FakeQdrantPoint("m2", {"attention_salience": 0.4, "narrative_importance": "medium"}),
+            FakeQdrantPoint("m3", {"attention_salience": 0.8, "narrative_importance": "high"}),
+        ]
+        fake = FakeQdrant(points)
+        monkeypatch.setattr("modules.consciousness.qdrant", fake)
+        monkeypatch.setattr("modules.workspace.qdrant", fake, raising=False)
+
+        from modules.consciousness import apply_salience_decay
+        result = apply_salience_decay(decay_rate=0.05)
+
+        # Medium memories should have been decayed
+        assert len(fake.payloads_set) == 2, f"Expected 2 set_payload calls, got {len(fake.payloads_set)}"
+        # Check the new values
+        for call in fake.payloads_set:
+            new_sal = call["payload"]["attention_salience"]
+            point_id = call["points"][0]
+            if point_id == "m1":
+                assert abs(new_sal - 0.55) < 0.01
+            elif point_id == "m2":
+                assert abs(new_sal - 0.35) < 0.01
+
+    def test_salience_decay_respects_floor(self, monkeypatch):
+        """Salience never drops below 0.1."""
+        points = [
+            FakeQdrantPoint("m1", {"attention_salience": 0.12, "narrative_importance": "low"}),
+        ]
+        fake = FakeQdrant(points)
+        monkeypatch.setattr("modules.consciousness.qdrant", fake)
+        monkeypatch.setattr("modules.workspace.qdrant", fake, raising=False)
+
+        from modules.consciousness import apply_salience_decay
+        apply_salience_decay(decay_rate=0.05)
+
+        assert len(fake.payloads_set) == 1
+        new_sal = fake.payloads_set[0]["payload"]["attention_salience"]
+        assert new_sal >= 0.1, f"Salience {new_sal} below floor 0.1"
+
+    def test_emotional_decay_toward_baseline(self, monkeypatch, clean_pad):
+        """PAD moves toward mood baseline without overshoot."""
+        from modules.config import _emotional_state
+
+        # Set extreme state
+        _emotional_state['current'] = {
+            'pleasure': 0.8, 'arousal': 0.9, 'dominance': 0.7,
+            'timestamp': datetime.now(TZ_COL).isoformat(),
+            'trigger': 'extreme_test',
+        }
+        _emotional_state['mood'] = {
+            'pleasure': 0.2, 'arousal': 0.1, 'dominance': 0.3,
+            'last_updated': None,
+        }
+        _emotional_state['decay_rate'] = 0.1
+        _emotional_state['history'] = []
+
+        # Mock classify/text functions to avoid qdrant deps
+        monkeypatch.setattr("modules.consciousness._classify_emotion", lambda p, a, d: "neutral")
+        monkeypatch.setattr("modules.consciousness._get_emotion_text", lambda l: "neutral")
+
+        from modules.consciousness import apply_emotional_decay
+        result_json = apply_emotional_decay()
+        result = json.loads(result_json)
+
+        # After one decay: current should move 10% toward baseline
+        new_p = result["current"]["pleasure"]
+        new_a = result["current"]["arousal"]
+        new_d = result["current"]["dominance"]
+
+        # Expected: 0.8 + (0.2 - 0.8) * 0.1 = 0.74
+        assert abs(new_p - 0.74) < 0.01, f"pleasure: expected 0.74, got {new_p}"
+        # Expected: 0.9 + (0.1 - 0.9) * 0.1 = 0.82
+        assert abs(new_a - 0.82) < 0.01, f"arousal: expected 0.82, got {new_a}"
+        # Expected: 0.7 + (0.3 - 0.7) * 0.1 = 0.66
+        assert abs(new_d - 0.66) < 0.01, f"dominance: expected 0.66, got {new_d}"
+
+        # No overshoot: new values still further from baseline than baseline
+        assert new_p > 0.2
+        assert new_a > 0.1
+        assert new_d > 0.3
+
+
+# ============================================================
+# BUCKET 5: Prospective Memory Persistence
+# Ref: McDaniel & Einstein 2007, Burgess et al. 2011
+# ============================================================
+
+class TestProspectiveMemory:
+    """Bucket 5: activation decay, floor, and cross-session persistence."""
+
+    def test_activation_floor_respected(self, monkeypatch):
+        """_compute_activation with 100h decay >= PM_ACTIVATION_FLOOR[critical] - tolerance."""
+        # Mock noise to 0 for determinism
+        monkeypatch.setattr("modules.prospective.random.gauss", lambda mu, sigma: 0)
+
+        from modules.prospective import _compute_activation, PM_ACTIVATION_FLOOR
+
+        result = _compute_activation(
+            stored_activation=0.95,
+            priority="critical",
+            hours_since_last_maintenance=100.0,
+            partial_match_count=0,
+            expiry=None,
+            now=None,
+        )
+
+        floor = PM_ACTIVATION_FLOOR["critical"]
+        assert result >= floor - 0.01, \
+            f"Activation {result} below floor {floor} for critical after 100h"
+
+    def test_deadline_urgency_boost(self, monkeypatch):
+        """Intention at 1 day > intention at 30 days (with noise=0)."""
+        monkeypatch.setattr("modules.prospective.random.gauss", lambda mu, sigma: 0)
+
+        from modules.prospective import _compute_activation
+
+        now = datetime.now(TZ_COL)
+        close_expiry = now + timedelta(days=1)
+        far_expiry = now + timedelta(days=30)
+
+        act_close = _compute_activation(
+            stored_activation=0.65,
+            priority="medium",
+            hours_since_last_maintenance=1.0,
+            expiry=close_expiry,
+            now=now,
+        )
+
+        act_far = _compute_activation(
+            stored_activation=0.65,
+            priority="medium",
+            hours_since_last_maintenance=1.0,
+            expiry=far_expiry,
+            now=now,
+        )
+
+        assert act_close > act_far, \
+            f"Close deadline ({act_close}) should > far deadline ({act_far})"
+
+    def test_intentions_survive_roundtrip(self, continuity_db, mock_external_deps, monkeypatch):
+        """checkpoint -> JSON -> bridge -> critical intentions appear in bridge_text."""
+        from modules.session_bridge import checkpoint_session_close, load_session_bridge
+
+        critical_intentions = [
+            {"id": "c1", "action": "Deploy trading bot to production",
+             "priority": "critical", "activation": 0.95, "expiry": None},
+        ]
+        monkeypatch.setattr(
+            "modules.prospective.get_pending_intentions",
+            lambda limit=10: critical_intentions
+        )
+
+        checkpoint_session_close(session_summary="roundtrip test", source="flush")
+
+        # Reset intentions mock for load (bridge reads from checkpoint JSON, not live)
+        monkeypatch.setattr(
+            "modules.prospective.get_pending_intentions",
+            lambda limit=10: []
+        )
+
+        bridge = load_session_bridge()
+        assert bridge is not None
+        assert "Deploy trading bot" in bridge["bridge_text"]
+
+
+# ============================================================
+# BUCKET 6: Autobiographical Continuity
+# Ref: Conway & Pleydell-Pearce 2000, Conway 2005
+# ============================================================
+
+class TestAutobiographicalContinuity:
+    """Bucket 6: bridge text respects length limits and temporal grounding."""
+
+    def test_bridge_text_max_chars(self, continuity_db, mock_external_deps, monkeypatch):
+        """Bridge with very long data <= BRIDGE_MAX_CHARS (500)."""
+        from modules.session_bridge import checkpoint_session_close, load_session_bridge, BRIDGE_MAX_CHARS
+
+        long_summary = "A" * 500
+        long_intentions = [
+            {"id": f"lg{i}", "action": f"Very long action description number {i} " + "Z" * 100,
+             "priority": "high", "activation": 0.8, "expiry": None}
+            for i in range(5)
+        ]
+        monkeypatch.setattr(
+            "modules.prospective.get_pending_intentions",
+            lambda limit=10: long_intentions
+        )
+
+        checkpoint_session_close(
+            session_summary=long_summary,
+            decisions="Big decision " * 50,
+            goal_stack=[{"goal": "big", "status": "active", "next_step": "N" * 200}],
+            source="flush",
+        )
+
+        bridge = load_session_bridge()
+        assert bridge is not None
+        assert len(bridge["bridge_text"]) <= BRIDGE_MAX_CHARS, \
+            f"Bridge text {len(bridge['bridge_text'])} chars > {BRIDGE_MAX_CHARS}"
+
+    def test_bridge_includes_temporal_marker(self, continuity_db, mock_external_deps):
+        """Bridge always contains 'Han pasado' + 'horas'."""
+        from modules.session_bridge import checkpoint_session_close, load_session_bridge
+
+        checkpoint_session_close(session_summary="temporal marker test", source="flush")
+
+        bridge = load_session_bridge()
+        assert bridge is not None
+        assert "Han pasado" in bridge["bridge_text"]
+        assert "horas" in bridge["bridge_text"]
+
+
+# ============================================================
+# BUCKET 7: State-Dependent Retrieval
+# Ref: Tulving & Thomson 1973, Godden & Baddeley 1975
+# ============================================================
+
+class TestStateDependentRetrieval:
+    """Bucket 7: despertar_codi restores PAD from bridge (thin integration)."""
+
+    def _setup_despertar_mocks(self, monkeypatch, continuity_db, pad_p=0.6, pad_a=-0.3, pad_d=0.7):
+        """Set up the minimal mocks for despertar_codi."""
+        from modules.config import now_col as _now_col
+
+        # Insert a checkpoint with known PAD values
+        _insert_checkpoint(
+            continuity_db,
+            session_summary="Previous session with known PAD",
+            pad_pleasure=pad_p,
+            pad_arousal=pad_a,
+            pad_dominance=pad_d,
+            pad_trigger="known_trigger",
+            active_project="consciencia",
+        )
+
+        # Mock 1: _verificar_salud_memoria_interna -> healthy
+        monkeypatch.setattr(
+            "modules.consciousness._verificar_salud_memoria_interna",
+            lambda: {"ok": True, "total_memories": 100, "message": "OK"}
+        )
+
+        # Mock 2: _load_triggers -> empty (imported locally from modules.triggers)
+        monkeypatch.setattr(
+            "modules.triggers._load_triggers",
+            lambda: []
+        )
+
+        # Mock 3: _verificar_tareas_vencidas -> empty (imported locally from modules.maintenance)
+        monkeypatch.setattr(
+            "modules.maintenance._verificar_tareas_vencidas",
+            lambda: []
+        )
+
+        # Mock 4: qdrant.scroll -> empty (no critical memories to display)
+        fake_qdrant = FakeQdrant([])
+        monkeypatch.setattr("modules.consciousness.qdrant", fake_qdrant)
+
+        # Mock 5: memory.search -> empty
+        mock_memory = MagicMock()
+        mock_memory.search.return_value = {"results": []}
+        monkeypatch.setattr("modules.consciousness.memory", mock_memory)
+
+        # Mock 6: load_session_state fallback -> None (imported locally from modules.flush)
+        monkeypatch.setattr("modules.flush.load_session_state", lambda: None)
+
+        # Mock 7: event_bus.emit -> no-op
+        from modules.events import event_bus
+        monkeypatch.setattr(event_bus, "emit", lambda *a, **kw: None)
+
+    def test_pad_restored_at_wakeup(self, continuity_db, mock_external_deps, monkeypatch, clean_pad):
+        """despertar_codi() restores PAD from bridge, not (0,0,0)."""
+        from modules.config import _emotional_state
+
+        self._setup_despertar_mocks(monkeypatch, continuity_db)
+
+        from modules.consciousness import despertar_codi
+        result_text = despertar_codi()
+
+        # PAD should be restored from checkpoint
+        current = _emotional_state['current']
+        assert abs(current['pleasure'] - 0.6) < 0.01, f"pleasure: {current['pleasure']}"
+        assert abs(current['arousal'] - (-0.3)) < 0.01, f"arousal: {current['arousal']}"
+        assert abs(current['dominance'] - 0.7) < 0.01, f"dominance: {current['dominance']}"
+
+    def test_restored_pad_trigger_marks_bridge(self, continuity_db, mock_external_deps, monkeypatch, clean_pad):
+        """Trigger indicates provenance ('restored_from_bridge')."""
+        from modules.config import _emotional_state
+
+        self._setup_despertar_mocks(monkeypatch, continuity_db)
+
+        from modules.consciousness import despertar_codi
+        result_text = despertar_codi()
+
+        assert "SESSION BRIDGE" in result_text
+        trigger = _emotional_state['current'].get('trigger', '')
+        assert "restored_from_bridge" in trigger, f"trigger: {trigger}"
+
+
+# ============================================================
+# BUCKET 8: Sleep Loop Mechanics
+# Ref: Diekelmann & Born 2010
+# ============================================================
+
+class TestSleepLoopMechanics:
+    """Bucket 8: tick ordering and report format cap."""
+
+    def test_ticks_execute_in_order(self, continuity_db, monkeypatch):
+        """Tick execution order == TICK_ORDER: [prospective, health, consolidation, homeostasis]."""
+        from modules.sleep_loop import TICK_ORDER
+
+        execution_order = []
+
+        def make_tick_mock(name):
+            def tick_fn(budget_ms):
+                execution_order.append(name)
+                return {"tick": name, "ok": True, "detail": "mocked", "elapsed_ms": 10}
+            return tick_fn
+
+        # Patch the tick dispatch functions
+        for tick_name in TICK_ORDER:
+            monkeypatch.setattr(
+                f"modules.sleep_loop._tick_{tick_name}",
+                make_tick_mock(tick_name)
+            )
+
+        from modules.sleep_loop import run_sleep_loop
+        result = run_sleep_loop(reason="test", budget_ms=8000)
+
+        assert execution_order == TICK_ORDER, \
+            f"Expected {TICK_ORDER}, got {execution_order}"
+
+    def test_sleep_report_format_capped(self):
+        """format_sleep_report() output <= 600 chars."""
+        from modules.sleep_loop import format_sleep_report
+
+        # Create tick results with long details
+        tick_results = [
+            {"tick": name, "ok": True, "detail": "D" * 200, "elapsed_ms": 1000}
+            for name in ["prospective", "health", "consolidation", "homeostasis"]
+        ]
+
+        report = format_sleep_report(tick_results, 4000, "test")
+        assert len(report) <= 600, f"Report {len(report)} chars > 600"
+
+
+# ============================================================
+# BUCKET 9: Source Priority
+# Ref: Johnson et al. 1993 (Source Monitoring Framework)
+# ============================================================
+
+class TestSourcePriority:
+    """Bucket 9: hook after flush -> deduped, flush data preserved."""
+
+    def test_hook_cannot_upgrade_flush(self, continuity_db, mock_external_deps):
+        """Hook after flush -> deduped=True, flush data preserved."""
+        from modules.session_bridge import checkpoint_session_close
+
+        r1 = checkpoint_session_close(
+            session_summary="flush data - high priority",
+            source="flush",
+        )
+        assert r1["ok"] is True
+
+        r2 = checkpoint_session_close(
+            session_summary="hook data - should be ignored",
+            source="hook",
+        )
+        assert r2.get("deduped") is True
+
+        # Verify flush data preserved
+        conn = sqlite3.connect(continuity_db)
+        row = conn.execute(
+            "SELECT session_summary, source FROM session_checkpoints "
+            "ORDER BY created_at DESC LIMIT 1"
+        ).fetchone()
+        conn.close()
+
+        assert row[0] == "flush data - high priority"
+        assert row[1] == "flush"
